% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_graduation.R
\name{add_all_subgroup}
\alias{add_all_subgroup}
\title{Add "all" subgroup to graduation data}
\usage{
add_all_subgroup(tidy_data)
}
\arguments{
\item{tidy_data}{Tidy graduation data}
}
\value{
Data frame with added "all" subgroup records
}
\description{
Adds a summary record that aggregates all diploma types.
}
\keyword{internal}
