% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_graduation.R
\name{calc_state_aggregates}
\alias{calc_state_aggregates}
\title{Calculate state aggregates (for v1 data)}
\usage{
calc_state_aggregates(tidy_data)
}
\arguments{
\item{tidy_data}{Tidy graduation data}
}
\value{
Data frame with added state aggregate records
}
\description{
For v1 era (2019-2022), calculates state-level aggregates by summing
all school-level records.
}
\keyword{internal}
