% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{clean_id}
\alias{clean_id}
\title{Clean ID values}
\usage{
clean_id(x)
}
\arguments{
\item{x}{Vector of ID values}
}
\value{
Character vector of cleaned IDs
}
\description{
Removes extra whitespace and ensures IDs are character type.
}
\keyword{internal}
