% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{clean_phone}
\alias{clean_phone}
\title{Clean phone number}
\usage{
clean_phone(x)
}
\arguments{
\item{x}{Vector of phone numbers}
}
\value{
Character vector of cleaned phone numbers
}
\description{
Standardizes phone number formatting.
}
\keyword{internal}
