% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{clean_principal_name}
\alias{clean_principal_name}
\title{Clean principal name}
\usage{
clean_principal_name(x)
}
\arguments{
\item{x}{Vector of principal names}
}
\value{
Character vector of cleaned names
}
\description{
Cleans up principal name formatting.
}
\keyword{internal}
