% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{create_empty_directory_df}
\alias{create_empty_directory_df}
\title{Create empty directory data frame}
\usage{
create_empty_directory_df()
}
\value{
Empty data frame with expected columns
}
\description{
Create empty directory data frame
}
\keyword{internal}
