% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{create_empty_division_df}
\alias{create_empty_division_df}
\title{Create empty division data frame}
\usage{
create_empty_division_df(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Empty data frame with expected columns
}
\description{
Create empty division data frame
}
\keyword{internal}
