% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_sqp_enrollment}
\alias{download_sqp_enrollment}
\title{Download enrollment from School Quality Profiles}
\usage{
download_sqp_enrollment(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
List with school and division data frames
}
\description{
Downloads enrollment data from VDOE's School Quality Profiles website.
The site provides CSV downloads of Fall Membership data.
}
\keyword{internal}
