% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_sqp_file}
\alias{download_sqp_file}
\title{Download a specific SQP enrollment file}
\usage{
download_sqp_file(end_year, level)
}
\arguments{
\item{end_year}{School year end}

\item{level}{"school" or "division"}
}
\value{
Data frame with enrollment data
}
\description{
Download a specific SQP enrollment file
}
\keyword{internal}
