% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_school_year}
\alias{format_school_year}
\title{Format school year string}
\usage{
format_school_year(end_year)
}
\arguments{
\item{end_year}{School year end (e.g., 2024)}
}
\value{
Character string in format "2023-24"
}
\description{
Converts an end year to a school year string.
}
\keyword{internal}
