% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_school_year_long}
\alias{format_school_year_long}
\title{Format long school year string}
\usage{
format_school_year_long(end_year)
}
\arguments{
\item{end_year}{School year end (e.g., 2024)}
}
\value{
Character string in format "2023-2024"
}
\description{
Converts an end year to a full school year string.
}
\keyword{internal}
