% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_available_years}
\alias{get_available_years}
\title{Get available years of Virginia enrollment data}
\usage{
get_available_years()
}
\value{
Integer vector of available years (2016-2025)
}
\description{
Returns a vector of years for which enrollment data is available
from the Virginia Department of Education (VDOE).
}
\details{
VDOE provides Fall Membership enrollment data through the School Quality
Profiles website. Historical data is available from approximately 2016
to the present.

Note: VDOE typically releases Fall Membership data in late fall/early winter
following the school year start. For example, 2024-25 (end_year=2025) data
is usually available by December 2024.
}
\examples{
get_available_years()

# Check the current range
range(get_available_years())
}
