% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{get_dir_cache_path}
\alias{get_dir_cache_path}
\title{Get directory cache file path}
\usage{
get_dir_cache_path(type)
}
\arguments{
\item{type}{Cache type ("directory_tidy" or "directory_raw")}
}
\value{
Full path to cache file
}
\description{
Get directory cache file path
}
\keyword{internal}
