% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_graduation.R
\name{get_graduation_url}
\alias{get_graduation_url}
\title{Get graduation data download URL}
\usage{
get_graduation_url(end_year)
}
\arguments{
\item{end_year}{School year end (e.g., 2023 for 2022-23 school year)}
}
\value{
Character URL to CSV file
}
\description{
Returns the CKAN download URL for a given graduation year.
}
\keyword{internal}
