% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{get_raw_directory}
\alias{get_raw_directory}
\title{Download raw school directory data from Virginia DOE}
\usage{
get_raw_directory()
}
\value{
Data frame with raw directory data
}
\description{
Downloads school directory data from the Virginia Department of Education's
directory website.
}
\keyword{internal}
