% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_raw_enr}
\alias{get_raw_enr}
\title{Download raw enrollment data from Virginia DOE}
\usage{
get_raw_enr(end_year)
}
\arguments{
\item{end_year}{School year end (e.g., 2024 for 2023-24 school year)}
}
\value{
List with school and division data frames
}
\description{
Downloads school and division enrollment data from the Virginia Department
of Education's School Quality Profiles website.
}
\keyword{internal}
