% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_graduation.R
\name{get_raw_graduation}
\alias{get_raw_graduation}
\title{Download raw graduation rate data}
\usage{
get_raw_graduation(end_year, cache_dir = get_cache_dir())
}
\arguments{
\item{end_year}{School year end (e.g., 2023 for 2022-23 school year)}

\item{cache_dir}{Directory to cache downloaded files}
}
\value{
Data frame with raw graduation data as provided by VDOE
}
\description{
Downloads graduation rate CSV from VDOE Open Data Portal and returns it as
a data frame with minimal processing.
}
\keyword{internal}
