% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_grade_code}
\alias{map_grade_code}
\title{Map grade codes to standard format}
\usage{
map_grade_code(grade_code)
}
\arguments{
\item{grade_code}{Character or numeric grade code}
}
\value{
Character grade in standard format (PK, K, 01-12)
}
\description{
Converts VDOE grade codes to standardized format.
}
\keyword{internal}
