% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_vdoe_race}
\alias{map_vdoe_race}
\title{Map VDOE race/ethnicity codes to standard names}
\usage{
map_vdoe_race(race_name)
}
\arguments{
\item{race_name}{Character race/ethnicity from VDOE data}
}
\value{
Character standardized race name
}
\description{
Converts VDOE race/ethnicity categories to standardized names.
}
\keyword{internal}
