% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_division_enr}
\alias{process_division_enr}
\title{Process division-level enrollment data}
\usage{
process_division_enr(division_data, end_year)
}
\arguments{
\item{division_data}{Data frame with division enrollment from VDOE}

\item{end_year}{School year end}
}
\value{
Processed division data frame
}
\description{
Process division-level enrollment data
}
\keyword{internal}
