% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{read_dir_cache}
\alias{read_dir_cache}
\title{Read directory data from cache}
\usage{
read_dir_cache(type)
}
\arguments{
\item{type}{Cache type ("directory_tidy" or "directory_raw")}
}
\value{
Cached data frame
}
\description{
Read directory data from cache
}
\keyword{internal}
